Allow the checkstyle binary to be configured
This allows users to select the version of checkstyle that they
want to use without being rail-roaded into taking the version
that ships with `rules_contrib_jvm`.

We also remove the `config_file` option, and instead always make
users specify the `checkstyle_config` that they want to use. This
will likely reduce boilerplate and duplication in people's build
files, and also makes it easier for us to figure out the right
binary to use.

Closes #10