Only visit each ClassOrInterfaceType once
For large repetitive classes, e.g. generated protobuf or avro files
which have a great many methods which interact with the same handful of
types, repeatedly processing the same types over and over again is very
expensive.

For instance, this drops the time taken to process
https://github.com/harness/harness-core/blob/develop/280-batch-processing/src/generated/java/io/harness/avro/ClusterBillingData.java
which is generated code containing a great many references to just 16
types down from 24 seconds to 4 seconds.

4 seconds is still a lot; we should probably do some parallelisation
here, but 4 seconds is a lot better than 24 seconds (particularly when
we currently have a hard cut-off of 30 seconds before we give up and
crash).