Fix up helper file deps
This now adds deps on actual targets, via the resolver, and also ensures
that the helper library is always generated where it should be.

There's a significant question here for module-mode, as to whether we
should be generating one `java_library` for all of the code, or a
test-specific `java_library` and a production `java_library. Currently,
this code generates a single shared one, as generating separate ones
would require non-trivial rework of how imports are resolved.