Deduplicate and sort imports (#55)
* Add SortedSet data structure

There are a bunch of places in this code that we:
1. Pass around slices assuming they're set-like, but then append things
   to them that they may already contain.
2. Manually deduplicate slices via a map.
3. `append` to slices to make copies of them, then re-use the
   pre-appended slice.

This data structure will allow us to track the semantics of
already-sorted-deduplicated values, as well as forcing more explicit
`Clone` operations where underlying buffers should not be re-used.

* Deduplicate and sort imports

This avoids generating a bunch of duplicated imports. It also fixes up
the fact that we were previously leaking packages across java_test
targets in module mode (as evidenced by the removals of the
non-TODO-annotated lines in
java/gazelle/testdata/java_test_per_file/src/test/java/com/example/modulewithhelpers/BUILD.want).