Reduce generate to a single codepath
Before, code was duplicated across the module/non-module and
test-per-file/test-per-suite axes, which led to inconsistencies and
made it hard to modify the behaviour across the axes. It was also hard
to reason about what the essential differences were across the axes,
because completely independent code was responsible for generation in
each.

Instead, have a single codepath, with conditional logic to conditionally
accumulate the differences in common containers, and express what to
generate as a result of those common containers.

Also:
* Make more clear to what root assorted paths are relative.
* Use parameterised factory functions for each rule generation.
* Set testonly on all test helper libraries, not just in non-module
  mode.
* Don't set broad visibility for test helper libraries.
* Properly set testonly packages on non-module mode test helper
  libraries.
* Generate helper libraries, and helper-only test suites, if helpers but
  no tests are present.

Note that in module mode, java_binary targets are now generated before
java_library targets, which is consistent with the behaviour in
non-module mode. This accounts for most of the test churn.