Split proto generation into a separate java package (#57)
We currently perform package-level not class-level resolution, so the
existing test data would actually fail when hitting the resolver because
we wouldn't be able to identify the proto as the source of this package.

Instead, use a unique package name (as we'd currently recommend users
do). We may change this expectation in the future, but for now, this is
a constraint of the system.