Add `junit-platform-suite` to runner filters
This is to allow
[@Suite](https://junit.org/junit5/docs/current/api/org.junit.platform.suite.api/org/junit/platform/suite/api/Suite.html)
annotations to be honored when users need to leverage the
[JUnit Platform Suite Engine](https://junit.org/junit5/docs/current/user-guide/#junit-platform-suite-engine)
(as long as they pass the necessary runtime dependencies, of course).

Adding `junit-platform-suite` to the runner filters allows in particular
that classes listed by `@SelectClasses` get actually taken into account
(among other features of the JUnit Platform Suite Engine).