Suppress warnings about security manager on java 17 (#90)
Have another crack at sorting out the warning printed by
the JRE when using Java 17+ and running a junit5 test.

The previous fix prevented a warning being displayed at
runtime, but then caused the JRE to emit a warning that
the `SecurityManager` had been set on the command line.

To avoid that, we now avoid setting the `SecurityManager`
on the command line, and we're back to setting it in
code.

To avoid a warning being displayed, we now "simply"
check the version of Java we're using at runtime. If it's
Java 11, then we use the `SecurityManager` as we did in
the first place. On Java 17, however, we make a series
of poorly considered choices that tie us to the OpenJDK
to reach into the internals of `System` and tell it that
we've already registered a `SecurityManager` instance.
This prevents the warning message from being displayed.

On the off-chance that we're running on a JRE that has
already had `SecurityManager` removed (I hear that may
be as soon as Java 20), we provide a fall-back in the
shape of a no-op control that does nothing.

All this to prevent a test calling `System.exit`....