gazelle: Generate exports for non-private return types (#158)
Return values which are used by calling code require the return type to
be on the classpath in order to compile uses of that value. Accordingly,
we should export those dependencies, to make sure the depender has them
on the classpath while compiling, to avoid strict deps issues.

We don't do this for private functions, but all other visibilities may
be called from other targets, so may need exporting.