fix(testresult): add skipped message to <skipped> not <failure>
Currently skipped tests with messages return
`<skipped></skipped><failure message="...">`, but all the other examples
of JUnit I can find return `<skipped message="...">`.

The inclusion of both a `<skipped>` and `<failure>` confuses parsers,
making them think that the test failed when it was actually skipped.

Skipping tests was added in
https://github.com/bazel-contrib/rules_jvm/pull/37.