Rename format pipeline
Our CI checks aren't currently required to merge.

As far as I can tell, this name key is the only thing you can use to
describe whether a check is required, and the fact that it matches the
name for the other GitHub action means we can't make both required.