Fix the JUnit 5 XML output (#246)
The output for the Junit5 XML did not match the output from the Gradle
(our gold standard) or Maven (a secondary goal). This took digging
further into the JUnit internal test setup and deciphering how each test
is processed, and pushing it back into the form we want it in. It would
be nice if JUnit itself did this but ...