Rename IsTestPath to IsTestPackage (#271)
This is actually operating on package names, and we only ever give it
package names.

This is sigificant because package names always have / as path
separators, whereas on Windows, paths use \ as path separators.

This code should always be operating on /s. The code already assumed as
much, so be more clear in the code and tests around it that this is what
we're doing.