refactor: remove runner attribute from create_jvm_test_suite (#289)
In `create_jvm_test_suite.bzl`, the `runner` attribute is not utilised
within the `define_test` function. By the time we reach the
`create_jvm_test_suite` macro, this attribute is redundant because the
specific test runner can be directly specified within the `define_test`
function provided by the user. So, this PR removes references to the
runner from the macro.