Optimize junit dependencies for each test target (#303)
This change simplifies the dependency list for the junit suite test
targets by creating a export-only library used to group the dependencies
together via an export-only java_library. The primary reason for this is
to simplify the project model for IntelliJ using the new Bazel plugin.
With this change, the IntelliJ model does not duplicate the dependencies
for each test class in the Aspects and the internal project model and
instead only refers to the java library.