[Checkstyle] use better internal names
WIth the addition of #283, the checkstyle execution script added two internal names. The problem is either overlap with the config file name, the checkstyle fails with a stack trace.

Use a name for these internal temp files less likely to collide (and error) with the user supplied config files.

This fixes #311.