Ensure deps referenced in java_test_suite make vars are passed to tests (#332)
https://github.com/bazel-contrib/rules_jvm/pull/303 introduced a
regression where if a `java_test_suite` references a dep in its
jvm_flags via `$(location)` or other Make variable, it will not be
present in the deps of the generated `java_test` leading to analysis
errors. This PR adds an example demonstrating the issue and suggests a
possible fix.

Fixes #334