[ci] Disable embedded javaparser integration tests on Windows (#379)
The Windows CI builds are failing. Rolling back history to test with
historical commits (where the same CI run is known to have passed)
reveals that there must be some kind of change in the environment, as
those historical runs also fail. The only tests that fail are to do with
the embedded javaparser.

There are two clues about what's going on in the logs, and the results
sometimes differ:

1. There are problems loading `java/lang/Object`, which hints at some
problem with the JVM being used. Looking at the java version on the
`PATH`, it _looks_ like the correct java will be used, but perhaps
something like a strict action env is selecting the wrong version?
2. Occasionally there are problems writing a file because there's a
transient lock on it. Extensive checking around this suggests this isn't
the actual cause.

Since only the embedded javaparser tests are failing, we now mark all of
them as `skip-windows`. A follow up PR, once someone can replicate the
build issue locally, should re-enable them. By disabling these tests for
now, other PRs in this repo can continue to land.