[gazelle] Add support for resources in gradle sourcesets (#370)
This means that if we see a directory structure like:

  src/foo/java
  src/foo/resources

We'll now generate a target in the `resources` directory which will be
included in the target generated in the `java` directory. This models
the behaviour of both maven and gradle which will put those resources
into the single jar that they build.