[gazelle] Add support for Kotlin sources (#375)
This adds support for generating build files from Kotlin sources, as
well as mixed Kotlin/Java sources. Language features which add implicit
dependencies are supported via `exports` from the libraries that create
them.

As with the existing java parser, the new `KtParser` relies on the
tooling provided by JetBrains to accurately parse the source files.

A follow up PR is required to properly support test generation, because
at the moment this will always generate a `java_test_suite` with Kotlin
sources. For now, a "fix" for this is to override the `java_test_suite`
with a custom mapping (eg. `# gazelle:map_kind java_test_suite
kt_jvm_test_suite //kotlin:defs.bzl`) though it is clearly not ideal.

The initial work and heavy lifting was by Morgan Roff, but I've since
continued working on this PR.

---------

Co-authored-by: MorganR <morgan@roff.dev>
Co-authored-by: Morgan Roff <roffmorgan@gmail.com>