Expose portions of the java_junit5_test macro logic via //java:defs.bzl (#382)
# Summary

Take portions of the `java_junit5_test` macro and expose it as a method
and constant via `//java:defs.bzl`.

## Problem

The `java_junit5_test` macro has some internal logic for preparing
`java_test` to work with `JUnit5Runner`. Creating a test runner that
wraps `JUnit5Runner` would have to duplicate that logic in order to
expose the correct functionality.

## Solution

Take the logic out of the macro and place it into a function that can be
called by any macro when creating the `java_test` target.

## Testing

Added test cases to ensure the logic is available to be imported and
nominally returns the correct value. The test does not cover all
branches of the jvm_flags logic for `java_junit5_test`.