[junit] Use a javaagent to prevent `System.exit` being called (#396)
We've been accumulating clever hacks to try and deal with this, but when
we're using Java 25 there's no `SecurityManager`, and there wasn't a
great way to prevent tests prematurely exiting.
    
We now use a javaagent, which rewrites the `System.exit` to throw the
same `SecurityException` we used to use in the original Java 11 code
path. The nice thing with this approach is that the same approach can be
used by all versions of the JRE, simplifying our code.
    
We ensure that the test classpath is left unpolluted by making sure that
we load the ASM library in an isolated classloader. The agent classpath
shouldn't pollute the runtime classpath so things should be safe.